WIP: MemoryECInstanceBase updates parent instance offsets for supporting structs when updating struct array offsets BF: MemoryECInstanceBase::UpdateStructArrayOffsets uses _GetBaseObjectSize() to calculate address of current instance

+++ iModelCore/ecobjects/PublicApi/ECObjects/StandaloneECInstance.h
@@ -64,7 +64,7 @@ struct PerPropertyFlagsHolder
 struct ParentInstanceUnion
     {
     MemoryECInstanceBase const *    parentInstance;
-    Int64                           offset;
+    size_t                          offset;
     };
 
 /*=================================================================================**//**
